Extract serializer / deserializer
