+++ pinecone/openapi_support/serializer.py
@@ -1,0 +1,50 @@
+import io
+
+from typing import Any
+
+
+from .exceptions import PineconeApiValueError
+from .model_utils import ModelNormal, ModelSimple, ModelComposed, date, datetime, model_to_dict
+
+
+class Serializer:
+    @staticmethod
+    def get_file_data_and_close_file(file_instance: io.IOBase) -> bytes:
+        file_data = file_instance.read()
+        file_instance.close()
+        return file_data
+
+    @classmethod
+    def sanitize_for_serialization(cls, obj) -> Any:
+        """Prepares data for transmission before it is sent with the rest client
+        If obj is None, return None.
+        If obj is str, int, long, float, bool, return directly.
+        If obj is datetime.datetime, datetime.date
+            convert to string in iso8601 format.
+        If obj is list, sanitize each element in the list.
+        If obj is dict, return the dict.
+        If obj is OpenAPI model, return the properties dict.
+        If obj is io.IOBase, return the bytes
+        :param obj: The data to serialize.
+        :return: The serialized form of data.
+        """
+        if isinstance(obj, (ModelNormal, ModelComposed)):
+            return {
+                key: cls.sanitize_for_serialization(val)
+                for key, val in model_to_dict(obj, serialize=True).items()
+            }
+        elif isinstance(obj, io.IOBase):
+            return cls.get_file_data_and_close_file(obj)
+        elif isinstance(obj, (str, int, float, bool)) or obj is None:
+            return obj
+        elif isinstance(obj, (datetime, date)):
+            return obj.isoformat()
+        elif isinstance(obj, ModelSimple):
+            return cls.sanitize_for_serialization(obj.value)
+        elif isinstance(obj, (list, tuple)):
+            return [cls.sanitize_for_serialization(item) for item in obj]
+        if isinstance(obj, dict):
+            return {key: cls.sanitize_for_serialization(val) for key, val in obj.items()}
+        raise PineconeApiValueError(
+            "Unable to prepare type {} for serialization".format(obj.__class__.__name__)
+        )
